fixup! Move autoupdate version resolving logic in a separate package
